+++ validity/src/Data/Validity.hs
@@ -61,6 +61,8 @@ module Data.Validity
     , Validation(..)
     , ValidationChain(..)
     , checkValidity
+    , validationIsValid
+    , prettyValidate
     , prettyValidation
     -- * Re-exports
     , Monoid(..)
@@ -547,12 +549,30 @@ checkValidity a =
         Validation [] -> Right a
         Validation errs -> Left errs
 
--- | validate a given value, and return a nice error if the value is invalid.
-prettyValidation :: Validity a => a -> Either String a
-prettyValidation a =
-    case checkValidity a of
-        Right a_ -> Right a_
-        Left errs -> Left $ intercalate "\n" $ map (errCascade . toStrings) errs
+-- | Check if a 'Validation' concerns a valid value.
+validationIsValid :: Validation -> Bool
+validationIsValid v = case v of
+    Validation [] -> True
+    _ -> False
+
+-- | Validate a given value
+--
+-- This function will return a nice error if the value is invalid.
+-- It will return the original value in 'Right' if it was valid,
+-- as evidence that it has been validated.
+prettyValidate :: Validity a => a -> Either String a
+prettyValidate a = case prettyValidation $ validate a of
+    Just e -> Left e
+    Nothing -> Right a
+
+-- | Render a `Validation` in a somewhat pretty way.
+--
+-- This function will return 'Nothing' if the 'Validation' concerned a valid value.
+prettyValidation :: Validation -> Maybe String
+prettyValidation v =
+    case v of
+        Validation [] -> Nothing
+        Validation errs -> Just $ intercalate "\n" $ map (errCascade . toStrings) errs
   where
     toStrings (Violated s) = ["Violated: " ++ s]
     toStrings (Location s vc) = s : toStrings vc