+++ validity/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## [Unreleased]
 
+### Added
+
+- `prettyValidate`, `validationIsValid`, `prettyValidation`
+
+### Changed
+
+- Renamed `prettyValidation` to `prettyValidate` before adding the new `prettyValidation`.
+
 ## [0.8.0.0] - 2018-08-25
 
 ### Added